requirements.txt
django
djangorestframework
djoser
django-filter
django-cors-headers
mysqlclient
